Continue working on abilities

+++ lab_03/inc/tui.h
@@ -14,7 +14,7 @@ int execute_operation(SparseMatrix **matrix, SparseVector **vector);
 
 size_t input_enum(size_t max_options, char **options);
 
-size_t input_value(char *title, bool min_limit, bool max_limit, int min_value, int max_value);
+int input_value(char *title, bool min_limit, bool max_limit, int min_value, int max_value);
 
 void input_matrix(SparseMatrix *matrix);
